feat: 로또 수동 생성 기능 추가
- 구매 가격을 Money 객체로 wrapping

+++ src/main/java/lotto/domain/RankResult.java
@@ -16,7 +16,7 @@ public class RankResult {
         return count;
     }
 
-    public Integer getPrizePerRank() {
+    public Long getPrizePerRank() {
         return rank.getPrize() * count;
     }
